Worker/backup/pyproject.toml
[tool.poetry]
name = "backup"
version = "0.1.0"
description = ""
authors = ["Lucas <lucas.mgu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyDrive = "^1.3.1"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
